requirements.txt
# Core dependencies
dspy-ai>=2.5.0
sqlalchemy>=2.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
duckduckgo-search>=6.0.0
mcp>=1.0.0
openai>=1.0.0
anthropic>=0.34.0
requests>=2.31.0
tenacity>=8.0.0
pyyaml>=6.0.0

# Semantic similarity
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# Development dependencies
mypy>=1.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
pre-commit>=3.6.0
types-requests>=2.31.0
types-pyyaml>=6.0.0
